Add insert purchase controller endpoint
It was supposed to make the operation as a Transaction, but
multidocument transactions are only supported on replica-set or sharded
cluster.
A subscription() on the unreturned flux is necesary for persisting the
information, otherwise is not saved.
When a purchase is done it's supposed to substract the quantity sold on
each of the products.